refactor: update metadata handling and logger integration
